Refactored the YieldInstrumenter into its own file
